+++ src/common/interfaces/nest-application.interface.ts
@@ -1,12 +1,12 @@
 import { MicroserviceConfiguration } from '@nestjs/microservices';
 import { INestMicroservice } from './index';
-import { SocketIoAdapter } from './socket-io-adapter.interface';
+import { WebSocketAdapter } from './web-socket-adapter.interface';
 
 export interface INestApplication {
     init(): void;
     listen(port: number, callback?: () => void);
     setGlobalPrefix(prefix: string): void;
-    setIoAdapter(adapter: SocketIoAdapter): void;
+    useWebSocketAdapter(adapter: WebSocketAdapter): void;
     connectMicroservice(config: MicroserviceConfiguration): INestMicroservice;
     getMicroservices(): INestMicroservice[];
     startAllMicroservices(callback: () => void): void;