+++ src/common/interfaces/socket-io-adapter.interface.ts
@@ -1,4 +1,0 @@
-export interface SocketIoAdapter {
-    create(port: number);
-    createWithNamespace(port: number, namespace: string);
-}